Change type of money related elements to decimal

+++ VS_Console_Boulangerie_2/Bakery.cs
@@ -4,19 +4,19 @@ namespace VS_Console_Boulangerie_Niv2
 {
     public class Bakery
     {
-        private float CashRegisterAmount { get; set; } = 0;
+        private decimal CashRegisterAmount { get; set; } = 0;
 
         public void SellBaguette()
         {
-            CashRegisterAmount += 1.1F;
+            CashRegisterAmount += 1.1m;
         }
 
         public void SellBread()
         {
-            CashRegisterAmount += 2.6F;
+            CashRegisterAmount += 2.6m;
         }
 
-        public float GetCashRegisterAmount()
+        public decimal GetCashRegisterAmount()
         {
             return CashRegisterAmount;
         }
